requirements.txt
annotated-types==0.6.0
boto3==1.34.34
botocore==1.34.34
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
jmespath==1.0.1
pydantic==2.6.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.0.7